+++ apps/backend/src/app/views/templates/EmailTemplate.tsx
@@ -96,7 +96,12 @@ export const EmailTemplate = ({
         </Text>
       )}
       <Text style={{ ...secondaryTextStyle, ...answerMargin }}>
-        {qa.answer}
+        {qa.answer.split(/\r?\n/).map((line, i, arr) => (
+          <React.Fragment key={i}>
+            {line}
+            {i < arr.length - 1 && <br />}
+          </React.Fragment>
+        ))}
       </Text>
     </>
   )