+++ apps/frontend/src/features/admin-form/responses/ResponsesPage/storage/worker/decryption.worker.ts
@@ -23,7 +23,6 @@ import { downloadAndDecryptAttachmentsAsZip } from '../utils/downloadAndDecryptA
 import {
   buildFormFieldMetaMap,
   processDecryptedContent,
-  processDecryptedContentV3,
   processDecryptedContentV4,
 } from '../utils/processDecryptedContent'
 
@@ -109,11 +108,9 @@ async function decryptSubmissionData(
   {
     submissionData,
     secretKey,
-    useV4,
   }: {
     submissionData: SubmissionStreamDto
     secretKey: string
-    useV4?: boolean
   },
 ): Promise<
   | {
@@ -146,57 +143,36 @@ async function decryptSubmissionData(
       break
     }
     case SubmissionType.Multirespondent: {
-      if (useV4) {
-        const formFieldsMeta = buildFormFieldMetaMap(submissionData.form_fields)
-        const decryptedV4 = ctx.formsgSdk.cryptoV3.decryptToV4(
-          secretKey,
-          {
-            encryptedSubmissionSecretKey:
-              submissionData.encryptedSubmissionSecretKey,
-            encryptedContent,
-            verifiedContent,
-            version,
-          },
-          formFieldsMeta,
-        )
-        if (decryptedV4) {
-          mrfSubmissionSecretKey = decryptedV4.submissionSecretKey
-          decryptedResponses = processDecryptedContentV4(
-            submissionData.form_fields,
-            decryptedV4.responses,
-            decryptedV4.verified,
-          )
-          break
-        }
-        datadogLogs.logger.warn(
-          'Could not decrypt MRF response in v4, falling back to v3',
-          { submissionId: submissionData._id },
-        )
-      }
-
-      // Perform v3 decryption if v4 deryption fails or is not attempted (useV4 = false)
-      const decryptedObject = ctx.formsgSdk.cryptoV3.decrypt(secretKey, {
-        encryptedSubmissionSecretKey:
-          submissionData.encryptedSubmissionSecretKey,
-        encryptedContent,
-        verifiedContent,
-        version,
-      })
-      if (!decryptedObject) {
-        datadogLogs.logger.error(
-          'Invalid decryption for multirespondent response',
-        )
+      const formFieldsMeta = buildFormFieldMetaMap(submissionData.form_fields)
+      const decryptedV4 = ctx.formsgSdk.cryptoV3.decryptToV4(
+        secretKey,
+        {
+          encryptedSubmissionSecretKey:
+            submissionData.encryptedSubmissionSecretKey,
+          encryptedContent,
+          verifiedContent,
+          version,
+        },
+        formFieldsMeta,
+      )
+      if (!decryptedV4) {
+        datadogLogs.logger.error('Could not decrypt MRF response in v4', {
+          submissionId: submissionData._id,
+        })
         return {
           isSubmissionDecryptionSuccessful: false,
         }
       }
-      mrfSubmissionSecretKey = decryptedObject.submissionSecretKey
-      decryptedResponses = await processDecryptedContentV3(
+
+      mrfSubmissionSecretKey = decryptedV4.submissionSecretKey
+      decryptedResponses = processDecryptedContentV4(
         submissionData.form_fields,
-        decryptedObject,
+        decryptedV4.responses,
+        decryptedV4.verified,
       )
       break
     }
+
     default: {
       // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const _: never = submissionData
@@ -445,10 +421,9 @@ async function _parseAndDecryptSubmissionData(
   {
     submissionStreamDtoString,
     secretKey,
-    useV4,
   }: Pick<
     SubmissionDataForDecryption,
-    'submissionStreamDtoString' | 'secretKey' | 'useV4'
+    'submissionStreamDtoString' | 'secretKey'
   >,
 ): Promise<DecryptionResult> {
   let submission: SubmissionStreamDto
@@ -468,7 +443,6 @@ async function _parseAndDecryptSubmissionData(
   const decryptSubmissionDataResult = await decryptSubmissionData(ctx, {
     submissionData: submission,
     secretKey,
-    useV4,
   })
 
   if (!decryptSubmissionDataResult.isSubmissionDecryptionSuccessful) {