+++ apps/frontend/src/features/form/constants.ts
@@ -1,9 +1,3 @@
-import { BasicField } from 'formsg-shared/types'
-
-export const NON_RESPONSE_FIELD_SET = new Set([
-  BasicField.Section,
-  BasicField.Statement,
-  BasicField.Image,
-])
+export { NON_RESPONSE_FIELD_SET } from 'formsg-shared/utils/field'
 
 export const SAVE_DRAFT_INDEXEDDB_STORE_NAME = 'savedDrafts'