+++ apps/frontend/src/features/workspace/components/CreateFormModal/CreateFormModalContent/CreateFormDetailsScreen.tsx
@@ -1,7 +1,8 @@
-import { Controller, RegisterOptions } from 'react-hook-form'
+import { Controller } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
 import { BiRightArrowAlt } from 'react-icons/bi'
 import {
+  Box,
   Container,
   FormControl,
   ModalBody,
@@ -13,27 +14,19 @@ import {
 import { FormResponseMode } from 'formsg-shared/types/form/form'
 
 import { GUIDE_PREVENT_EMAIL_BOUNCE } from '~constants/links'
-import { useFormTitleValidationRules } from '~utils/formValidation'
 import Button from '~components/Button'
 import FormErrorMessage from '~components/FormControl/FormErrorMessage'
-import FormFieldMessage from '~components/FormControl/FormFieldMessage'
 import FormLabel from '~components/FormControl/FormLabel'
 import InlineMessage from '~components/InlineMessage'
-import Input from '~components/Input'
 
 import DataClassificationInfoBox from '~features/admin-form/settings/components/DataClassificationInfoBox'
 
-import { WorkspaceRowsProvider } from '../../WorkspaceFormRow/WorkspaceRowsContext'
-import {
-  CreateFormWizardInputProps,
-  useCreateFormWizard,
-} from '../CreateFormWizardContext'
+import { useCreateFormWizard } from '../CreateFormWizardContext'
 
 import { EmailFormRecipientsInput } from './EmailFormRecipientsInput'
+import { EscapeHatchLink } from './EscapeHatchLink'
 import { FormResponseOptions } from './FormResponseOptions'
-
-/** The length of form title to start showing warning text */
-const FORM_TITLE_LENGTH_WARNING = 65
+import { FormTitleInput } from './FormTitleInput'
 
 const getTrackingSubmissionActionName = (
   responseModeValue: FormResponseMode,
@@ -62,23 +55,20 @@ export const CreateFormDetailsScreen = (): JSX.Element => {
     isLoading,
     isFetching,
     modalHeader,
-    isSingpass,
     hasMyInfoChildren,
+    isMrfCutoverEnabled,
+    goToStorageModeDetails,
   } = useCreateFormWizard()
   const {
-    register,
     control,
     formState: { errors },
     watch,
   } = formMethods
 
-  const titleInputValue = watch('title')
   const responseModeValue = watch('responseMode')
   const handleEmailButtonPress = () => {
     handleEmailFeedbackSubmit()
   }
-
-  const formTitleValidationRules = useFormTitleValidationRules()
 
   return (
     <>
@@ -89,91 +79,78 @@ export const CreateFormDetailsScreen = (): JSX.Element => {
       </ModalHeader>
       <ModalBody whiteSpace="pre-wrap">
         <Container maxW="45rem" p={0}>
-          <FormControl isRequired isInvalid={!!errors.title} mb="2.25rem">
-            <FormLabel useMarkdownForDescription>
-              {t('features.workspace.modals.forms.create.details.name.label')}
-            </FormLabel>
-            <Skeleton isLoaded={!isFetching}>
-              <Input
-                autoFocus
-                {...register(
-                  'title',
-                  formTitleValidationRules as RegisterOptions<
-                    CreateFormWizardInputProps,
-                    'title'
-                  >,
-                )}
-              />
-            </Skeleton>
-            <FormErrorMessage>{errors.title?.message}</FormErrorMessage>
-            {titleInputValue?.length > FORM_TITLE_LENGTH_WARNING ? (
-              <FormFieldMessage>
-                {t(
-                  'features.workspace.modals.forms.create.details.name.message',
-                )}
-              </FormFieldMessage>
-            ) : null}
-          </FormControl>
-          <FormControl isRequired isInvalid={!!errors.responseMode} mb="2.5rem">
-            <FormLabel
-              description={t(
-                'features.workspace.modals.forms.create.details.type.description',
-              )}
+          <FormTitleInput />
+          {isMrfCutoverEnabled ? (
+            <Box my="2.5rem">
+              <EscapeHatchLink onClick={goToStorageModeDetails} />
+            </Box>
+          ) : (
+            <FormControl
+              isRequired
+              isInvalid={!!errors.responseMode}
+              mb="2.5rem"
             >
-              {t('features.workspace.modals.forms.create.details.type.label')}
-            </FormLabel>
-            <Skeleton isLoaded={!isFetching}>
-              <Controller
-                name="responseMode"
-                control={control}
-                render={({ field }) => (
-                  <WorkspaceRowsProvider>
+              <FormLabel
+                description={t(
+                  'features.workspace.modals.forms.create.details.type.description',
+                )}
+              >
+                {t('features.workspace.modals.forms.create.details.type.label')}
+              </FormLabel>
+              <Skeleton isLoaded={!isFetching}>
+                <Controller
+                  name="responseMode"
+                  control={control}
+                  render={({ field }) => (
                     <FormResponseOptions
                       {...field}
                       hasMyInfoChildren={hasMyInfoChildren}
                       handleEmailButtonPress={handleEmailButtonPress}
                     />
-                  </WorkspaceRowsProvider>
-                )}
-                rules={{
-                  required: t(
-                    'features.workspace.modals.forms.create.errors.responseMode.required',
-                  ),
-                }}
-              />
-            </Skeleton>
-            <FormErrorMessage>{errors.responseMode?.message}</FormErrorMessage>
-            {hasMyInfoChildren && (
-              <InlineMessage mt="2rem">
-                {t(
-                  'features.workspace.modals.forms.create.errors.noMyInfoChildrenInMrf',
-                )}
-              </InlineMessage>
-            )}
-          </FormControl>
-          {(responseModeValue === FormResponseMode.Encrypt ||
-            responseModeValue === FormResponseMode.Email) && (
-            <FormControl
-              isRequired={responseModeValue === FormResponseMode.Email}
-              isInvalid={!!errors.emails}
-              mb="2.25rem"
-            >
-              <FormLabel
-                isRequired={responseModeValue === FormResponseMode.Email}
-                useMarkdownForDescription
-                description={t(
-                  'features.workspace.modals.forms.create.details.notifications.description',
-                  { GUIDE_PREVENT_EMAIL_BOUNCE },
-                )}
-              >
-                {t(
-                  'features.workspace.modals.forms.create.details.notifications.label',
-                )}
-              </FormLabel>
-              <EmailFormRecipientsInput />
+                  )}
+                  rules={{
+                    required: t(
+                      'features.workspace.modals.forms.create.errors.responseMode.required',
+                    ),
+                  }}
+                />
+              </Skeleton>
+              <FormErrorMessage>
+                {errors.responseMode?.message}
+              </FormErrorMessage>
+              {hasMyInfoChildren && (
+                <InlineMessage mt="2rem">
+                  {t(
+                    'features.workspace.modals.forms.create.errors.noMyInfoChildrenInMrf',
+                  )}
+                </InlineMessage>
+              )}
             </FormControl>
           )}
-          <DataClassificationInfoBox />
+          {!isMrfCutoverEnabled &&
+            (responseModeValue === FormResponseMode.Encrypt ||
+              responseModeValue === FormResponseMode.Email) && (
+              <FormControl
+                isRequired={responseModeValue === FormResponseMode.Email}
+                isInvalid={!!errors.emails}
+                mb="2.25rem"
+              >
+                <FormLabel
+                  isRequired={responseModeValue === FormResponseMode.Email}
+                  useMarkdownForDescription
+                  description={t(
+                    'features.workspace.modals.forms.create.details.notifications.description',
+                    { GUIDE_PREVENT_EMAIL_BOUNCE },
+                  )}
+                >
+                  {t(
+                    'features.workspace.modals.forms.create.details.notifications.label',
+                  )}
+                </FormLabel>
+                <EmailFormRecipientsInput />
+              </FormControl>
+            )}
+          {!isMrfCutoverEnabled && <DataClassificationInfoBox />}
           <Button
             rightIcon={<BiRightArrowAlt fontSize="1.5rem" />}
             type="submit"