+++ apps/frontend/src/features/workspace/components/CreateFormModal/CreateFormModalContent/EscapeHatchLink.tsx
@@ -1,0 +1,29 @@
+import { Text } from '@chakra-ui/react'
+
+import { composeEscapeHatchCopy } from '~utils/escapeHatchCopy'
+import InlineMessage from '~components/InlineMessage'
+import Link from '~components/Link'
+
+import { useUser } from '~features/user/queries'
+
+interface EscapeHatchLinkProps {
+  onClick: () => void
+}
+
+export const EscapeHatchLink = ({
+  onClick,
+}: EscapeHatchLinkProps): JSX.Element => {
+  const { user } = useUser()
+  const { prefix, linkText, suffix } = composeEscapeHatchCopy(user?.betaFlags)
+  return (
+    <InlineMessage variant="info">
+      <Text>
+        {prefix}
+        <Link cursor="pointer" onClick={onClick}>
+          {linkText}
+        </Link>
+        {suffix}
+      </Text>
+    </InlineMessage>
+  )
+}